+++ src/wallpaper_config.rs
@@ -8,28 +8,28 @@ use std::{
 use color_eyre::{eyre::ensure, Result};
 use serde::Deserialize;
 
-use crate::output::Output;
+use crate::wallpaper_info::WallpaperInfo;
 
-#[derive(Deserialize)]
-pub struct OutputConfig {
+#[derive(Deserialize, PartialEq)]
+pub struct WallpaperConfig {
     #[serde(flatten)]
-    data: HashMap<String, Arc<Output>>,
+    data: HashMap<String, Arc<WallpaperInfo>>,
     #[serde(skip)]
-    default_config: Arc<Output>,
+    default_config: Arc<WallpaperInfo>,
     #[serde(skip)]
     pub path: PathBuf,
     #[serde(skip)]
     pub reloaded: bool,
 }
 
-impl OutputConfig {
+impl WallpaperConfig {
     pub fn new_from_path(path: &Path) -> Result<Self> {
         ensure!(path.exists(), "Configuration file {path:?} does not exists",);
         let mut config_manager: Self = toml::from_str(&fs::read_to_string(path)?)?;
         config_manager.default_config = config_manager
             .data
             .get("default")
-            .unwrap_or(&Arc::new(Output::default()))
+            .unwrap_or(&Arc::new(WallpaperInfo::default()))
             .clone();
         for (name, config) in &config_manager.data {
             let path = config.path.as_ref().unwrap();
@@ -39,7 +39,8 @@ impl OutputConfig {
             );
             ensure!(
                 config.duration.is_none() || path.is_dir(),
-                "Duration can only be set when path points to a directory, for input {name}"
+                "for input '{name}', `path` is set to an image but `duration` is also set.
+Either remove `duration` or set `path` to a directory"
             );
         }
 
@@ -48,7 +49,7 @@ impl OutputConfig {
         Ok(config_manager)
     }
 
-    pub fn get_output_by_name(&self, name: &str) -> Arc<Output> {
+    pub fn get_output_by_name(&self, name: &str) -> Arc<WallpaperInfo> {
         self.data.get(name).unwrap_or(&self.default_config).clone()
     }
 }